# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Ghana Super App (Sankofa) API
  version: "1.0.0"
  description: >
    Maintainability follow-up: this is the first machine-readable description
    of the API surface — previously it lived only in the zod schemas and
    route files themselves. Coverage is intentionally uneven right now:

      - Auth, Rides and Payments (the original core loop, and the modules
        with the deepest test coverage) are fully specified: request/response
        bodies, status codes, auth requirements.
      - Every other module (Admin, and the eight service tiles — Eat, Fix,
        Utilities, Liquidity, Fleet, Property, Travel, plus Drivers,
        Vehicles, Locations, Notifications, Pricing, Ratings, Support, Users)
        is listed with its real path, method, tag and auth requirement, but
        without full request/response schemas yet — filling those in is real,
        traceable work (see the technical-readiness checklist's
        "Maintainability" section), not a placeholder to hide.

    All endpoints are under the `/api/v1` base path shown in `servers`.
    Every response follows one envelope: `{ "success": true, "data": ... }`
    on success, `{ "success": false, "error": { "code", "message" } }` on
    failure — this is enforced by errorHandler.ts, not repeated on every
    operation below.
servers:
  - url: /api/v1
    description: Base path for every route below (see main.ts)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Short-lived access token from POST /auth/verify-otp or
        POST /auth/refresh. Required on every route below except
        /auth/request-otp, /auth/verify-otp, /auth/refresh and
        /payments/webhook (which authenticates by HMAC signature instead —
        see the payments.webhook path and ADR 0002).
...
security:
  - bearerAuth: []
paths:
...
  /rides/{id}:
    get:
      tags: [Rides]
      summary: Get one ride by id
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Ride returned }
        "404": { description: Not found }
  /rides/{id}/cancel:
    post:
      tags: [Rides]
      summary: Cancel a ride (customer or driver, depending on status)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Ride cancelled } }
  /rides/{id}/arrived:
    post:
      tags: [Rides]
      summary: Driver marks arrival at pickup
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Ride status updated to ARRIVED } }
  /rides/{id}/start:
    post:
      tags: [Rides]
      summary: Driver starts the trip
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Ride status updated to IN_PROGRESS } }
  /rides/{id}/complete:
    post:
      tags: [Rides]
      summary: Driver completes the trip
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Ride status updated to COMPLETED, payment created } }

  # ------------------------------------------------------------------
  # Payments — fully specified (webhook fraud fix, ADR 0002)
  # ------------------------------------------------------------------
...
  /eat/restaurants/{id}:
    get:
      tags: [Eat]
      summary: Get one restaurant + menu
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
...
  /eat/orders/{id}:
    get:
      tags: [Eat]
      summary: Get one order
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /eat/orders/{id}/status:
    patch:
      tags: [Eat]
      summary: Update order status (restaurant/rider side)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
...
  /fleet/vehicles/{id}/assign:
    post:
      tags: [Fleet]
      summary: Assign a driver to a fleet vehicle
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /fleet/vehicles/{id}/unassign:
    post:
      tags: [Fleet]
      summary: Unassign a driver from a fleet vehicle
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
...
  /travel/bookings/{id}:
    get:
      tags: [Travel]
      summary: Get one booking
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /travel/bookings/{id}/cancel:
    post:
      tags: [Travel]
      summary: Cancel a booking
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }

  # Admin — every route requires PLATFORM_ADMIN (adminRouter.use(requireRole(...)))
...